requirements.txt
tensorflow>=2.5
pillow>=8.0
numpy>=1.19